Enhance PWA manifest and offline support
