+++ root/frontend/public/manifest.webmanifest
@@ -1,0 +1,77 @@
+{
+  "name": "Экосистема ГУУ — цифровой кампус",
+  "short_name": "Экосистема ГУУ",
+  "description": "Расписание пар, новости университета и интерактивная карта кампуса — всё доступно даже офлайн.",
+  "lang": "ru",
+  "dir": "ltr",
+  "display": "standalone",
+  "display_override": ["standalone", "browser"],
+  "start_url": "/",
+  "scope": "/",
+  "orientation": "portrait-primary",
+  "theme_color": "#0b63f4",
+  "background_color": "#f4f8fd",
+  "categories": ["education", "productivity"],
+  "shortcuts": [
+    {
+      "name": "Расписание",
+      "description": "Быстрый переход к расписанию занятий",
+      "url": "/schedule",
+      "icons": [
+        {
+          "src": "/maskable-icon-192.png",
+          "sizes": "192x192",
+          "type": "image/png"
+        }
+      ]
+    },
+    {
+      "name": "Новости",
+      "description": "Свежие новости кампуса",
+      "url": "/news",
+      "icons": [
+        {
+          "src": "/maskable-icon-192.png",
+          "sizes": "192x192",
+          "type": "image/png"
+        }
+      ]
+    },
+    {
+      "name": "Карта кампуса",
+      "description": "Маршруты и корпуса на карте ГУУ",
+      "url": "/map",
+      "icons": [
+        {
+          "src": "/maskable-icon-192.png",
+          "sizes": "192x192",
+          "type": "image/png"
+        }
+      ]
+    }
+  ],
+  "icons": [
+    {
+      "src": "/guu_logo.png",
+      "sizes": "192x192",
+      "type": "image/png"
+    },
+    {
+      "src": "/maskable-icon-192.png",
+      "sizes": "192x192",
+      "type": "image/png",
+      "purpose": "maskable any"
+    },
+    {
+      "src": "/guu_logo.png",
+      "sizes": "512x512",
+      "type": "image/png"
+    },
+    {
+      "src": "/maskable-icon-512.png",
+      "sizes": "512x512",
+      "type": "image/png",
+      "purpose": "maskable any"
+    }
+  ]
+}
