+++ root/frontend/public/offline.html
@@ -3,7 +3,7 @@
   <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1, viewport-fit=cover" />
-    <title>Экосистема ГУУ — офлайн</title>
+    <title>Экосистема ГУУ — офлайн режим</title>
     <style>
       :root {
         color-scheme: light dark;
@@ -16,6 +16,7 @@
         place-items: center;
         background: #f4f8fd;
         color: #0b0d11;
+        padding: 1rem;
       }
       @media (prefers-color-scheme: dark) {
         body {
@@ -24,7 +25,7 @@
         }
       }
       main {
-        max-width: 32rem;
+        max-width: 36rem;
         padding: 2.5rem 2rem;
         text-align: center;
       }
@@ -35,28 +36,57 @@
       p {
         margin: 0 auto 1.5rem;
         line-height: 1.6;
-        opacity: 0.8;
+        opacity: 0.85;
       }
-      a {
-        color: inherit;
-        text-decoration: none;
-        font-weight: 600;
-        border-bottom: 1px solid currentColor;
+      ul {
+        list-style: none;
+        padding: 0;
+        margin: 0 0 1.5rem;
+        display: grid;
+        gap: 0.5rem;
+      }
+      li {
+        padding: 0.75rem 1rem;
+        border-radius: 1rem;
+        background: rgba(11, 99, 244, 0.08);
+      }
+      @media (prefers-color-scheme: dark) {
+        li {
+          background: rgba(76, 141, 255, 0.16);
+        }
       }
       button {
         font: inherit;
-        padding: 0.75rem 1.5rem;
+        padding: 0.75rem 1.75rem;
         border-radius: 999px;
         border: none;
         cursor: pointer;
         background: #0b63f4;
         color: #fff;
+        transition: transform 0.2s ease, box-shadow 0.2s ease;
+      }
+      button:focus-visible {
+        outline: 2px solid currentColor;
+        outline-offset: 3px;
+      }
+      button:hover {
+        transform: translateY(-1px);
+        box-shadow: 0 10px 25px rgba(11, 99, 244, 0.3);
       }
       @media (prefers-color-scheme: dark) {
         button {
           background: #4c8dff;
           color: #0b0d11;
         }
+        button:hover {
+          box-shadow: 0 10px 25px rgba(76, 141, 255, 0.25);
+        }
+      }
+      small {
+        display: block;
+        margin-top: 2rem;
+        font-size: 0.875rem;
+        opacity: 0.65;
       }
     </style>
   </head>
@@ -64,13 +94,15 @@
     <main>
       <h1>Нет подключения к сети</h1>
       <p>
-        Похоже, что вы офлайн. Проверьте подключение к интернету и
-        попробуйте снова открыть приложение.
+        Экосистема ГУУ продолжит синхронизировать расписание, новости и карту кампуса,
+        как только появится интернет.
       </p>
+      <ul>
+        <li>Расписание и новости, просмотренные ранее, останутся доступными офлайн.</li>
+        <li>Интерактивная карта загрузится автоматически после восстановления связи.</li>
+      </ul>
       <button type="button" onclick="window.location.reload()">Повторить попытку</button>
-      <p style="margin-top:2rem; font-size:0.875rem; opacity:0.6;">
-        Экосистема ГУУ продолжит работать, как только появится соединение.
-      </p>
+      <small>Проверьте подключение или попробуйте открыть приложение позднее.</small>
     </main>
   </body>
 </html>
