+++ root/frontend/tests/e2e/offline.spec.ts
@@ -1,0 +1,38 @@
+import { expect, test } from "@playwright/test";
+import { useMockApi } from "./utils/mockApi";
+
+test.describe("PWA offline support", () => {
+  test("shows offline fallback page when network is unavailable", async ({ page, context }) => {
+    const mock = await useMockApi(page);
+    await mock.login(page);
+
+    await page.waitForLoadState("networkidle");
+    await page.evaluate(async () => {
+      await navigator.serviceWorker?.ready;
+    });
+
+    const hasController = await page.evaluate(() => Boolean(navigator.serviceWorker?.controller));
+    if (!hasController) {
+      await page.reload({ waitUntil: "networkidle" });
+      await page.evaluate(async () => {
+        await navigator.serviceWorker?.ready;
+      });
+    }
+
+    await page.goto("/offline.html", { waitUntil: "domcontentloaded" });
+    await expect(page.getByRole("heading", { name: "Нет подключения к сети" })).toBeVisible();
+
+    await page.goto("/dashboard", { waitUntil: "networkidle" });
+
+    await context.setOffline(true);
+    try {
+      await page.goto("/news", { waitUntil: "domcontentloaded" });
+      await expect(page.getByRole("heading", { name: "Нет подключения к сети" })).toBeVisible();
+      await expect(
+        page.getByText("Расписание и новости, просмотренные ранее, останутся доступными офлайн.")
+      ).toBeVisible();
+    } finally {
+      await context.setOffline(false);
+    }
+  });
+});
